feat : MemberCoupon 기능 구현 및 주문 상태 관련 기능 추가, CartItem 관련 버그 수정

+++ src/main/java/cart/persistence/cart/CartItemRowMapper.java
@@ -1,5 +1,6 @@
 package cart.persistence.cart;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -19,7 +20,7 @@ public class CartItemRowMapper implements RowMapper<CartItem> {
 	private Product mapProduct(ResultSet rs) throws SQLException {
 		Long productId = rs.getLong("product.id");
 		String name = rs.getString("name");
-		Long price = rs.getLong("price");
+		BigDecimal price = rs.getBigDecimal("price");
 		String imageUrl = rs.getString("image_url");
 		return new Product(productId, name, price, imageUrl);
 	}
